Extend ':has()' style invalidation support for pseudo classes.
This CL extends ':has()' invalidation support for user-action pseudo
classes and input value states pseudo classes in ':has()' argument.

- user-action
  - https://www.w3.org/TR/selectors-4/#useraction-pseudos
  - :hover, :active, :focus, :focus-within, :focus-visible

- input value states
  - https://www.w3.org/TR/selectors-4/#input-value-states
  - :checked, :indeterminate

To prevent invalidation triggered by the unnecessary pseudo state
change, the pseudo types of pseudos in ':has()' argument are
collected and stored in a HashSet.

When there is a pseudo state change mutation on an element marked as
AncestorsAffectedByHas, StyleEngine will match the mutation with the
HashSet to determine whether the mutation is related to the ':has()'
state change or not.

To prevent frequent invalidation triggered by some pseudo state change
such as ':hover' or ':focus' on an unnecessary element, the computed
style of the element affected by ':has()' have flags for each pseudos
(e.g. 'AncestorsAffectedByHoverInHas') which is inherited by its
children. The flags indicate that we need to walkup ancestors to
invalidate the 'AffectedByHas' elements when there is the pseudo state
change on the element. Similar with the other flags for ':has()', the
flags are set when while selector matching.

To prevent unnecessary ancestor invalidation while element insertion or
removal, a computed style extra field (AffectedByDescendantPseudoState)
was also added to indicate that the element can be affected by any
pseudo state change of a descendant element. If there is a style rule
using any pseudo in ':has()' argument, StyleEngine will check the extra
flag of ancestors when it failed to match the changed element with id/
class/attribute/tagname value set.

Bug: 669058
Change-Id: Ie726bf7108e991eca060a48c4a7a1b01713bfffe