Disallow container queries at or inside multicol containers.
Container queries will remain disallowed until we have full NG block
fragmentation support, for all layout types, including printing.

The reason is that, with container queries, style is recalculated during
layout. We cannot risk that we switch between legacy and NG layout while
we're there, since we may have started NG layout on ancestors that style
recalc suddenly wants to be legacy.

Had to improve the multicol detection mechanism that's run during style
recalc because of this, to ensure that we only trigger legacy fallback
when an element is actually going to become a multicol container (which
isn't the case for e.g. <span style="columns:2;">). Otherwise we may
trigger unnecessary legacy layout fallback on ancestors all the way up
to the block formatting context root, which may already be in the middle
of NG layout, if we're evaluating container queries.

Regarding the tests included: The inline-with-columns-* ones would only
crash without LayoutNGBlockFragmentation enabled, while all the others
would only crash *with* LayoutNGBlockFragmentation enabled.

Had to update a couple of unit tests that no longer get legacy fallback
(because there aren't actually any multicols there).

Bug: 1276898
AX-Relnotes: n/a
Change-Id: Ia75fe86d91966233ffeea43a3940113b6a431075