URLPattern: Emit `{}` around custom names followed by a regexp group.
This CL fixes another problem case from:

https://github.com/WICG/urlpattern/issues/145

In this case we need to preserve `{ ... }` braces around a named group
followed by a group that may be emitted in regexp `( ... )` group.  We
don't want `{:foo}(.*)` to be changed to `:foo(.*)` since that changes
the meaning from two groups to a single group.

Bug: 1263673
Change-Id: Ifc073d7bfb10f437180393b34cc3bf3c869d7f68